fix(openclaw): share one reading of an unexaminable vault
The OpenClaw runtime registers its own status tool with an independent
payload, and it carried the last instance of the conflation this wave has
been closing: an existence check that answers false for a vault that is
absent and for one that merely cannot be examined, so an unreadable
parent directory reported as an install that was never set up.

Rather than a second copy of the fix, the probe and the one formulation
of the condition move to core, where both runtimes consume them. Two
independent payloads publishing the same field cannot afford two
descriptions of the same failure.

Its query tool refused either way, so it was never falsely healthy, but
the reason was wrong in the same way the primary runtime's was: it told
an operator to create a vault they already have.
